split instance services into service + adapter pairs
Break the circular import cycle (runtime → instances → mixed module →
runtime) by splitting each instance service into:

- service.ts: Service class, layer, types, schemas (no runtime import)
- index.ts: thin adapter with runPromiseInstance wrappers

instances.ts now imports only service modules. Each service layer has
.pipe(Layer.fresh) at the definition site, and Layer.orDie is also
applied at the definition for FileWatcher and FileTime.

Also removes redundant Layer.fresh from withServices test fixture.
